+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Remove docs only relevant for the Exposition-CRD
 
+### Fixed
+- [#117] Add missing namespace to the maintenance-mode and dogu-starting middleware.
+
 ## [v6.1.1] - 2026-07-15
 ### Fixed
 - [#114] Create separate Ingress resources when dogu has different exposed services with same port